cea708: Fix location of the caption window placement
In the 708 specification, the origin for references to window location
is the top left corner (see CTA-708E Sec 8.2), but VLC's internal
representation has the origin at the bottom left corner of the
screen.  Hence we need to do a bit of math to handle this conversion
or else windows which are supposed to be at the bottom according to
the spec are rendered at the top.

In other words, a Y-value of 99% should be rendered at the bottom of
the screen, not at the top of the screen.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>