Write various VANC related stats to a file which can be picked up by marquee
Grab various stats related to CC, AFD, SCTE-104, and the SDI video
mode, and write them to files in a format where they can be picked
up by the marquee video filter and displayed on-screen.

The following command can be used for testing:

CARDINDEX=0
./bin/vlc-osx-static --sub-filter="marq{file=/tmp/vlc_${CARDINDEX}_line1.txt,x=10,y=10,position=-1,size=20}:marq{file=/tmp/vlc_${CARDINDEX}_line2.txt,x=10,y=30,position=-1,size=20}:marq{file=/tmp/vlc_${CARDINDEX}_line3.txt,x=10,y=50,position=-1,size=20}:marq{file=/tmp/vlc_${CARDINDEX}_line4.txt,x=10,y=70,position=-1,size=20}" decklink:// --decklink-card-index $CARDINDEX --decklink-tenbits 1

The decklink card index is used in the filenames to allow for
running multiple instances of VLC each capturing from a different
decklink device.